requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# HTTP Client
httpx==0.25.1
aiohttp>=3.8.0
requests==2.31.0

# Data Processing
pandas>=2.0.0,<2.1.0
numpy>=1.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# Caching
redis==5.0.1

# API Rate Limiting
limits==3.6.0

# Data Validation
pydantic==2.5.2
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# Utilities
beautifulsoup4>=4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Date/Time
pytz==2023.3.post1

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0

# Additional Packages
tenacity>=8.0.1